fix(port): USB port had wrong variable type causing timeout after 72 minutes
The timeout variable type was `uint32_t` instead of `int32_t`, causing the
timeout to be calculated as a negative number after 72 minutes. This
caused the port to stop working after 72 minutes.

+++ port/esp32_usb_cdc_acm_port.c
@@ -245,7 +245,7 @@ static void esp32_usb_start_timer(esp_loader_port_t *port, uint32_t ms)
 static uint32_t esp32_usb_remaining_time(esp_loader_port_t *port)
 {
     esp32_usb_cdc_acm_port_t *p = container_of(port, esp32_usb_cdc_acm_port_t, port);
-    int64_t remaining = ((int64_t)p->_time_end - esp_timer_get_time()) / 1000;
+    int64_t remaining = (p->_time_end - esp_timer_get_time()) / 1000;
     return (remaining > 0) ? (uint32_t)remaining : 0;
 }
